fix: refactor benchmarks
